nidaqmx example update
