+++ examples/nidaqmx_analog_input/measurement.py
@@ -11,10 +11,13 @@ import nidaqmx
 from _helpers import (
     ServiceOptions,
     configure_logging,
+    create_session_management_client,
+    get_grpc_device_channel,
     get_service_options,
     grpc_device_options,
     verbosity_option,
 )
+from _nidaqmx_helpers import _create_nidaqmx_task
 from nidaqmx.constants import TaskMode
 
 import ni_measurementlink_service as nims
@@ -46,12 +49,7 @@ def measure(pin_name, sample_rate, number_of_samples):
         sample_rate,
         number_of_samples,
     )
-    session_management_client = nims.session_management.Client(
-        grpc_channel=measurement_service.get_channel(
-            provided_interface=nims.session_management.GRPC_SERVICE_INTERFACE_NAME,
-            service_class=nims.session_management.GRPC_SERVICE_CLASS,
-        )
-    )
+    session_management_client = create_session_management_client(measurement_service)
     with contextlib.ExitStack() as stack:
         reservation = stack.enter_context(
             session_management_client.reserve_sessions(
@@ -80,7 +78,8 @@ def measure(pin_name, sample_rate, number_of_samples):
 
         measurement_service.context.add_cancel_callback(cancel_callback)
 
-        task = stack.enter_context(_create_nidaqmx_task(session_info))
+        grpc_device_channel = get_grpc_device_channel(measurement_service, nidaqmx)
+        task = stack.enter_context(_create_nidaqmx_task(session_info, grpc_device_channel))
         if not session_info.session_exists:
             task.ai_channels.add_ai_voltage_chan(session_info.channel_list)
 
@@ -96,31 +95,6 @@ def measure(pin_name, sample_rate, number_of_samples):
     _log_measured_values(voltage_values)
     logging.info("Completed measurement")
     return (voltage_values,)
-
-
-def _create_nidaqmx_task(
-    session_info: nims.session_management.SessionInformation,
-) -> nidaqmx.Task:
-    session_kwargs = {}
-    if service_options.use_grpc_device:
-        session_grpc_address = service_options.grpc_device_address
-
-        if not session_grpc_address:
-            session_grpc_channel = measurement_service.get_channel(
-                provided_interface=nidaqmx.GRPC_SERVICE_INTERFACE_NAME,
-                service_class="ni.measurementlink.v1.grpcdeviceserver",
-            )
-        else:
-            session_grpc_channel = measurement_service.channel_pool.get_channel(
-                target=session_grpc_address
-            )
-        session_kwargs["grpc_options"] = nidaqmx.GrpcSessionOptions(
-            session_grpc_channel,
-            session_name=session_info.session_name,
-            initialization_behavior=nidaqmx.SessionInitializationBehavior.AUTO,
-        )
-
-    return nidaqmx.Task(new_task_name=session_info.session_name, **session_kwargs)
 
 
 def _log_measured_values(samples, max_samples_to_display=5):
